Adds check for testdrive expiration
After 14 days (configurable in settings) the user should
see a styled message that tells them their testdrive has expired
and linking to our appsembler.com contact page if they want to
renew. The user will also see a message indicating how many days they
have left til the testdrive expires.

Once it's expired they will have access to their courses but in view
only mode...meaning they will not be able to access studio anymore to
edit the content of the course. The expiration date is tied to the
course creation date so that the user can invite more staff users
from their organization independently.

IMPORTANT: This is, however, predicated
on the fact that each user has only one course because we list all the
courses the user is enrolled in and pick the first one.

There is also a variable that can deny access to the course itself but
it's turned off for now since we do want them to have access to the
course.