Fixes days til expiration messages.
We pluralize "days" correctly and we make sure to display the default
message if the user doesn't have a course created.

Basically this should never happen (that a user doesn't have a course
created) but if it happens rather than saying "None days" it will print
"14 days" (or whatever the setting) because it assumes that the trial
has not started yet if the user does not have a course created.

This bug happened only if one registered directly on the edx site
without going through amc first. We should remove the ability for the
user to register directly on the edx side.