CMS - Search API to support Taxoman reindexing course search
Added package `search_api` to the CMS djangoapps. This package provides an HTTP POST method to start the course search reindex. This is added so that clients may use token based authentication instead of user/password authentication is is required for the existing course reindex functionality. Also, the exising course reindexing requires the user to be the course
author in order to reindex. This will not work for refreshing course search as the user is not expected to be the course author. Even though this django app is added to support the Taxoman API package, there is no dependency on taxoman_api.