add rough performance log/profiling to LMS dashboard view
Trying at a very high level to figure out where dashboard is spending
all its time on the mysterious slow requests.

This uses the existing logging mechanism and just generates some rough
perf data for each request to `/dashboard` so we can narrow it down.

This commit should be reverted once we have identified the
culprit (and probably replaced with something more surgical if we need
to narrow it down further).