Fixed the issue wth  missing Taxoman filters in the LMS course discovery sidebar
Facet values (terms) are required to be globally unique. This patch
fixes this by doing two things:

First, we add the facet value's model id to the facet name. This
compound value is then pushed to elastic search and retrieved by the
front end.

Second, we remove the trailing identifier from the facet name when
showing in the sidebar